increased mass measurement from kg to tonne

+++ Assets/Project/Source/Grids/Blocks/Components/Physical/PhysicalTemplatePartial.cs
@@ -12,7 +12,7 @@ namespace Exa.Grids.Blocks.Components
     {
         [SerializeField] private float maxHull;
         [SerializeField] private float armor;
-        [SerializeField] private short mass; // In kg
+        [SerializeField] private float mass; // In ton
 
         public float MaxHull => maxHull;
         public float Armor => armor;
